ldb: Add a generic tevent wrapper module
This wrapper module will be reused by both tdb and mdb back ends,
resulting in less code duplication.

It is meant as either augmenting the current module_ops for tevent-based
modules or even replacement for it. The advantage of the newer API is
that it encapsulated the tevent part, allowing for a nicer API -- the
modules only implement the operations.

Additionally, the ldb_request can now be made totally opaque and the
operations would only be able to access internals of operation-speficic
(ex: struct ldb_add for adding, struct ldb_modify for modification, ...)
structures, not the whole ldb_request.

If the tevent wrapper is used atop the current module_ops, then we will
save code duplication.